Update README.md
I am a curious and enthusiastic learner, driven by a desire to experience and accomplish as much as I can in life. Though I understand I can't do it all, I am determined to embrace every opportunity and dive into a variety of pursuits with an open mind and a sense of adventure.